+++ src/main/java/com/silviaferrari/pbs/model/Product.java
@@ -5,6 +5,8 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
+import java.math.BigDecimal;
+
 // indica che questa classe rappresenta una tabella di un database
 @Entity
 public class Product {
@@ -18,13 +20,13 @@ public class Product {
 
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private String imageUrl;
     //private int quantity;
 
     public Product() {}
 
-    public Product(String name, String description, double price, String imageUrl) {
+    public Product(String name, String description, BigDecimal price, String imageUrl) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -55,11 +57,11 @@ public class Product {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
